requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
argon2-cffi==23.1.0
aiohttp==3.8.6
